tabbing and file downloading changes

+++ plot1.R
@@ -8,17 +8,20 @@
 #clear contents 
 rm(list = ls())
 
+#Specify file name to be accessed, 
+fname <- "household_power_consumption.txt"
+
+if (!file.exists(fname)){
 #specify URL and download file
 file_url <- "https://d396qusza40orc.cloudfront.net/exdata%2Fdata%2Fhousehold_power_consumption.zip"
 dl_file <- download.file(file_url,destfile = "tmp")
 
 unzip("tmp")    #extract files
 unlink("tmp")   #delete file
+}
 
-#Specify file name to be accessed, 
 #then read and store in a variable
-file_name <- "household_power_consumption.txt"
-data <- read.table(file_name, header = TRUE, sep = ";", na.strings = "?")
+data <- read.table(fname, header = TRUE, sep = ";", na.strings = "?")
 
 #concatenate date and time, 
 #then convert to representable date and time
@@ -30,7 +33,7 @@ data$Date <- as.Date(data$Date, "%d/%m/%Y")
 #Store subset of data meeting the specified condition
 sub <- subset(data, Date >= as.Date("2007-02-01") & Date <= as.Date("2007-02-02"))
 
-#Plot 1
+# Plot 1
 hist(sub$Global_active_power, col = "red", xlab = "Global Active Power (kilowatts)", 
      ylab = "Frequency", main = "Global Active Power", breaks = 13, ylim = c(0,1300), xlim = c(0, 8))
